+++ bocadillo/recipes.py
@@ -1,10 +1,6 @@
-from typing import Any, Sequence
+from typing import Sequence
 
-
-from .meta import DocsMeta
-from .misc import overrides
-from .routing import HTTPRoute, RoutingMixin, WebSocketRoute
-from .templates import TemplatesMixin
+from .applications import App
 
 
 class RecipeBase:
@@ -14,7 +10,7 @@ class RecipeBase:
         assert prefix.startswith("/"), "recipe prefix must start with '/'"
         self.prefix = prefix
 
-    def __call__(self, app, root: str = ""):
+    def apply(self, app: App, root: str = ""):
         """Apply the recipe to an application.
 
         Should be implemented by subclasses.
@@ -26,7 +22,7 @@ class RecipeBase:
         raise NotImplementedError
 
 
-class Recipe(TemplatesMixin, RoutingMixin, RecipeBase, metaclass=DocsMeta):
+class Recipe(App):
     """A grouping of capabilities that can be merged back into an application.
 
     # Parameters
@@ -39,38 +35,27 @@ class Recipe(TemplatesMixin, RoutingMixin, RecipeBase, metaclass=DocsMeta):
     """
 
     def __init__(self, name: str, prefix: str = None, **kwargs):
+        super().__init__(name=name, **kwargs)
+
         if prefix is None:
             prefix = f"/{name}"
-        super().__init__(prefix=prefix, **kwargs)
+        assert prefix.startswith("/"), "recipe prefix must start with '/'"
+
+        self.prefix = prefix
         # DEPRECATED: 0.13.0
         self._templates_dir_given = "templates_dir" in kwargs
-        self.name = name
 
-    def get_template_globals(self):
-        # DEPRECATED: 0.13.0
-        return {"url_for": self.url_for}
+    def _get_own_url_for(self, name: str, **kwargs) -> str:
+        return self.prefix + super()._get_own_url_for(name, **kwargs)
 
-    @overrides(RoutingMixin.route)
-    def route(self, pattern: str, **kwargs) -> HTTPRoute:
-        kwargs["namespace"] = self.name
-        return super().route(self.prefix + pattern, **kwargs)
-
-    @overrides(RoutingMixin.websocket_route)
-    def websocket_route(self, pattern: str, **kwargs) -> WebSocketRoute:
-        return super().websocket_route(self.prefix + pattern, **kwargs)
-
-    def __call__(self, obj: Any, root: str = ""):
-        """Apply the recipe to an object.
-        
-        Said object must be a subclass `RoutingMixin` and `TemplateMixin`.
-        """
-        obj.http_router.mount(self.http_router, root=root)
-        obj.websocket_router.mount(self.websocket_router, root=root)
+    def apply(self, app: App, root: str = ""):
+        """Apply the recipe to an application."""
+        app.mount(prefix=root + self.prefix, app=self)
 
         # DEPRECATED: 0.13.0
         # Look for templates where the app does, if not specified
         if not self._templates_dir_given:
-            self.templates_dir = obj.templates_dir
+            self.templates_dir = app.templates_dir  # type: ignore
 
     @classmethod
     def book(cls, *recipes: "Recipe", prefix: str) -> "RecipeBook":
@@ -94,7 +79,7 @@ class RecipeBook(RecipeBase):
         super().__init__(prefix)
         self.recipes = recipes
 
-    def __call__(self, obj: Any, root: str = ""):
-        """Apply the recipe book to an object."""
+    def apply(self, app: App, root: str = ""):
+        """Apply the recipe book to an application."""
         for recipe in self.recipes:
-            recipe(obj, root=root + self.prefix)
+            recipe.apply(app, root=root + self.prefix)